Test build with MinGW on appveyor

+++ benchmarks/expand6_ginac.cpp
@@ -31,7 +31,7 @@ int main(int argc, char* argv[])
     a1 = symbol("a1");
     e = a0 + a1;
     f = 0;
-    for (int i = 2; i < N; i++) {
+    for (long long i = 2; i < N; i++) {
         std::ostringstream o;
         o << "a" << i;
         s = symbol(o.str());
